correct rebin function integer error

+++ knpy/utils/binning.py
@@ -21,6 +21,6 @@ def rebin(x, y, numbins):
         Rebinned bin heights.
     """
     assert len(x) % numbins == 0, "numbins has to be a factor of len(x)"
-    x_rebin = np.mean(x.reshape(len(x)/numbins, numbins), axis=1)
-    y_rebin = np.sum(y.reshape(len(x)/numbins, numbins), axis=1)
+    x_rebin = np.mean(x.reshape(int(len(x)/numbins), int(numbins)), axis=1)
+    y_rebin = np.sum(y.reshape(int(len(x)/numbins), int(numbins)), axis=1)
     return x_rebin, y_rebin
